requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
motor==3.3.2
pydantic==2.6.1
python-dotenv==1.0.0
pymongo==4.6.1
python-jose==3.3.0
passlib==1.7.4
emails==0.6.0
email-validator==2.1.0
bcrypt==4.1.2